+++ packages/core/src/bar-chart.ts
@@ -1,11 +1,12 @@
 // D3 Imports
 import { select, mouse } from "d3-selection";
-import { scaleBand } from "d3-scale";
+import { scaleBand, ScaleBand } from "d3-scale";
 import { min } from "d3-array";
 
 import { BaseAxisChart } from "./base-axis-chart";
 import { StackedBarChart } from "./stacked-bar-chart";
 import * as Configuration from "./configuration";
+import { ChartConfig, BarChartOptions, ChartType } from "./configuration";
 
 import { Tools } from "./tools";
 
@@ -40,7 +41,7 @@ const getMaxBarWidth = (maxWidth, currentBandWidth) => {
 	return maxWidth;
 };
 
-// returns true if the calculated bandwidth is greater than the maxWidth (if deinfed)
+// returns true if the calculated bandwidth is greater than the maxWidth (if defined)
 // i.e. if we should be constraining ourselves to a specific bar width
 const isWidthConstrained = (maxWidth, currentBandWidth) => {
 	if (!maxWidth) {
@@ -53,12 +54,11 @@ const isWidthConstrained = (maxWidth, currentBandWidth) => {
 };
 
 export class BarChart extends BaseAxisChart {
-	x: any;
-	x1?: any;
-	y: any;
-	colorScale: any;
+	x1?: ScaleBand<any>;
 
-	constructor(holder: Element, configs: any) {
+	options: BarChartOptions;
+
+	constructor(holder: Element, configs: ChartConfig<BarChartOptions>) {
 		// If this is a stacked bar chart, change the object prototype
 		if (configs.options.scales.y.stacked) {
 			if (getYMin(configs) >= 0) {
@@ -70,6 +70,13 @@ export class BarChart extends BaseAxisChart {
 
 		super(holder, configs);
 
+		// initialize options
+		if (configs.options) {
+			this.options = Tools.merge({}, Configuration.options.BAR, configs.options);
+		} else {
+			this.options = Tools.merge({}, Configuration.options.BAR);
+		}
+
 		// To be used for combo chart instances of a bar chart
 		const { axis } = configs.options;
 		if (axis) {
@@ -78,11 +85,11 @@ export class BarChart extends BaseAxisChart {
 			const width = chartSize.width - margins.left - margins.right;
 
 			this.x1 = scaleBand().rangeRound([0, width]).padding(Configuration.bars.spacing.bars);
-			this.x1.domain(configs.data.datasets.map(dataset => dataset.label))
+			this.x1.domain(this.data.datasets.map(dataset => dataset.label))
 				.rangeRound([0, getMaxBarWidth(Tools.getProperty(this.options, "bars", "maxWidth"), this.x.bandwidth())]);
 		}
 
-		this.options.type = "bar";
+		this.options.type = ChartType.BAR;
 	}
 
 	setXScale(xScale?: any) {
@@ -146,8 +153,8 @@ export class BarChart extends BaseAxisChart {
 						.attr("y", d => this.y(Math.max(0, d.value)))
 						.attr("width", this.x1.bandwidth())
 						.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
-						.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
-						.attr("stroke", d => this.options.accessibility ? this.colorScale[d.datasetLabel](d.label) : null)
+						.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
+						.attr("stroke", d => this.options.accessibility ? this.getStrokeColor(d.datasetLabel, d.label, d.value) : null)
 						.attr("stroke-width", Configuration.bars.default.strokeWidth)
 						.attr("stroke-opacity", d => this.options.accessibility ? 1 : 0);
 
@@ -193,9 +200,9 @@ export class BarChart extends BaseAxisChart {
 			.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
 			.style("opacity", 0)
 			.transition(this.getFillTransition())
-			.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
+			.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
 			.style("opacity", 1)
-			.attr("stroke", (d: any) => this.colorScale[d.datasetLabel](d.label))
+			.attr("stroke", (d: any) => this.getStrokeColor(d.datasetLabel, d.label, d.value))
 			.attr("stroke-width", Configuration.bars.default.strokeWidth);
 
 		addedBars.selectAll("rect.bar")
@@ -209,9 +216,9 @@ export class BarChart extends BaseAxisChart {
 			.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
 			.style("opacity", 0)
 			.transition(this.getFillTransition())
-			.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
+			.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
 			.style("opacity", 1)
-			.attr("stroke", (d: any) => this.colorScale[d.datasetLabel](d.label))
+			.attr("stroke", (d: any) => this.getStrokeColor(d.datasetLabel, d.label, d.value))
 			.attr("stroke-width", Configuration.bars.default.strokeWidth);
 
 		// Remove bar groups are no longer needed
@@ -237,11 +244,6 @@ export class BarChart extends BaseAxisChart {
 	}
 
 	updateElements(animate: boolean, rect?: any, g?: any) {
-		const { scales } = this.options;
-
-		const chartSize = this.getChartSize();
-		const height = chartSize.height - this.getBBox(".x.axis").height;
-
 		if (!rect) {
 			rect = this.innerWrap.selectAll("rect.bar");
 		}
@@ -262,8 +264,8 @@ export class BarChart extends BaseAxisChart {
 			.attr("y", d => this.y(Math.max(0, d.value)))
 			.attr("width", this.x1.bandwidth())
 			.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
-			.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
-			.attr("stroke", d => this.options.accessibility ? this.colorScale[d.datasetLabel](d.label) : null);
+			.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
+			.attr("stroke", d => this.options.accessibility ? this.getStrokeColor(d.datasetLabel, d.label, d.value) : null);
 	}
 
 	resizeChart() {
@@ -302,7 +304,7 @@ export class BarChart extends BaseAxisChart {
 			.on("mouseover", function(d) {
 				select(this)
 					.attr("stroke-width", Configuration.bars.mouseover.strokeWidth)
-					.attr("stroke", self.colorScale[d.datasetLabel](d.label))
+					.attr("stroke", self.getStrokeColor(d.datasetLabel, d.label, d.value))
 					.attr("stroke-opacity", Configuration.bars.mouseover.strokeOpacity);
 
 				self.showTooltip(d, this);
@@ -319,7 +321,7 @@ export class BarChart extends BaseAxisChart {
 				const { strokeWidth, strokeWidthAccessible } = Configuration.bars.mouseout;
 				select(this)
 					.attr("stroke-width", accessibility ? strokeWidthAccessible : strokeWidth)
-					.attr("stroke", accessibility ? self.colorScale[d.datasetLabel](d.label) : "none")
+					.attr("stroke", accessibility ? self.getStrokeColor(d.datasetLabel, d.label, d.value) : "none")
 					.attr("stroke-opacity", Configuration.bars.mouseout.strokeOpacity);
 
 				self.hideTooltip();