+++ packages/core/src/index.ts
@@ -8,11 +8,10 @@ import { DonutChart, DonutCenter } from "./donut-chart";
 import { BarChart } from "./bar-chart";
 import { LineChart } from "./line-chart";
 import { ComboChart } from "./combo-chart";
+import { ScatterChart } from "./scatter-chart";
 
 import * as Configuration from "./configuration";
 const defaultColors = Configuration.options.BASE.colors;
-
-import "./style.scss";
 
 export {
 	defaultColors,
@@ -23,5 +22,6 @@ export {
 	DonutCenter,
 	BarChart,
 	LineChart,
-	ComboChart
+	ComboChart,
+	ScatterChart
 };