+++ packages/core/src/line-chart.ts
@@ -2,46 +2,29 @@
 import { select, selectAll, mouse } from "d3-selection";
 import { line } from "d3-shape";
 
-import { BaseAxisChart } from "./base-axis-chart";
+import { ScatterChart } from "./scatter-chart";
 import * as Configuration from "./configuration";
+import { ChartConfig, LineChartOptions, ChartType } from "./configuration";
 
 import { getD3Curve } from "./services/curves";
+import { Tools } from "./tools";
 
-export class LineChart extends BaseAxisChart {
-	x: any;
-	y: any;
-
-	colorScale: any;
-
+export class LineChart extends ScatterChart {
 	lineGenerator: any;
 
-	constructor(holder: Element, configs: any) {
+	options: LineChartOptions;
+
+	constructor(holder: Element, configs: ChartConfig<LineChartOptions>) {
 		super(holder, configs);
 
-		this.options.type = "line";
-	}
+		// initialize options
+		if (configs.options) {
+			this.options = Tools.merge({}, Configuration.options.LINE, configs.options);
+		} else {
+			this.options = Tools.merge({}, Configuration.options.LINE);
+		}
 
-	getLegendType() {
-		return Configuration.legend.basedOn.SERIES;
-	}
-
-	addLabelsToDataPoints(d, index) {
-		const { labels } = this.displayData;
-
-		return d.data.map((datum, i) => ({
-			label: labels[i],
-			datasetLabel: d.label,
-			value: datum
-		}));
-	}
-
-	getCircleRadius() {
-		return this.options.points.radius || Configuration.charts.points.radius;
-	}
-
-	getCircleFill(radius, d) {
-		const circleShouldBeFilled = radius < Configuration.lines.points.minNonFilledRadius;
-		return circleShouldBeFilled ? this.colorScale[d.datasetLabel](d.label) : "white";
+		this.options.type = ChartType.LINE;
 	}
 
 	draw() {
@@ -49,14 +32,8 @@ export class LineChart extends BaseAxisChart {
 			.style("height", "100%");
 
 		const { line: margins } = Configuration.charts.margin;
-		const { scales } = this.options;
 
-		const chartSize = this.getChartSize();
-		const width = chartSize.width - margins.left - margins.right;
-		const height = chartSize.height - this.getBBox(".x.axis").height;
-
-		this.innerWrap.style("width", "100%")
-			.style("height", "100%");
+		this.innerWrap.style("width", "100%").style("height", "100%");
 
 		this.innerWrap.attr("transform", `translate(${margins.left}, ${margins.top})`);
 
@@ -85,28 +62,12 @@ export class LineChart extends BaseAxisChart {
 				.classed("lines", true);
 
 		gLines.append("path")
-			.attr("stroke", d => this.colorScale[d.label]())
+			.attr("stroke", d => this.getStrokeColor(d.label))
 			.datum(d => d.data)
 			.attr("class", "line")
 			.attr("d", this.lineGenerator);
 
-		const circleRadius = this.getCircleRadius();
-		gLines.selectAll("circle.dot")
-			.data((d, i) => this.addLabelsToDataPoints(d, i))
-			.enter()
-				.append("circle")
-				.attr("class", "dot")
-				.attr("cx", d => this.x(d.label) + this.x.step() / 2)
-				.attr("cy", d => this.y(d.value))
-				.attr("r", circleRadius)
-				.attr("fill", d => this.getCircleFill(circleRadius, d))
-				.attr("stroke", d => this.colorScale[d.datasetLabel](d.label));
-
-		// Hide the overlay
-		this.updateOverlay().hide();
-
-		// Dispatch the load event
-		this.dispatchEvent("load");
+		super.draw();
 	}
 
 	interpolateValues(newData: any) {
@@ -130,29 +91,13 @@ export class LineChart extends BaseAxisChart {
 			.classed("lines", true);
 
 		addedLineGroups.append("path")
-			.attr("stroke", d => this.colorScale[d.label]())
+			.attr("stroke", d => this.getStrokeColor(d.label))
 			.datum(d => d.data)
 			.style("opacity", 0)
 			.transition(this.getDefaultTransition())
 			.style("opacity", 1)
 			.attr("class", "line")
 			.attr("d", this.lineGenerator);
-
-		// Add line circles
-		const circleRadius = this.getCircleRadius();
-		addedLineGroups.selectAll("circle.dot")
-			.data((d, i) => this.addLabelsToDataPoints(d, i))
-			.enter()
-				.append("circle")
-				.attr("class", "dot")
-				.attr("cx", (d, i) => this.x(d.label) + this.x.step() / 2)
-				.attr("cy", (d: any) => this.y(d.value))
-				.attr("r", circleRadius)
-				.style("opacity", 0)
-				.transition(this.getDefaultTransition())
-				.style("opacity", 1)
-				.attr("fill", d => this.getCircleFill(circleRadius, d))
-				.attr("stroke", d => this.colorScale[d.datasetLabel](d.label));
 
 		// Remove lines that are no longer needed
 		gLines.exit()
@@ -161,22 +106,10 @@ export class LineChart extends BaseAxisChart {
 			.style("opacity", 0)
 			.remove();
 
-		// Add slice hover actions, and clear any slice borders present
-		this.addDataPointEventListener();
-
-		// Hide the overlay
-		this.updateOverlay().hide();
-
-		// Dispatch the update event
-		this.dispatchEvent("update");
+		super.interpolateValues(newData);
 	}
 
 	updateElements(animate: boolean, gLines?: any) {
-		const { scales } = this.options;
-
-		const chartSize = this.getChartSize();
-		const height = chartSize.height - this.getBBox(".x.axis").height;
-
 		if (!gLines) {
 			gLines = this.innerWrap.selectAll("g.lines");
 		}
@@ -193,103 +126,11 @@ export class LineChart extends BaseAxisChart {
 			.style("opacity", 1)
 			.attr("stroke", function(d) {
 				const parentDatum = select(this.parentNode).datum() as any;
-
-				return self.colorScale[parentDatum.label]();
+				return self.getStrokeColor(parentDatum.label);
 			})
 			.attr("class", "line")
 			.attr("d", this.lineGenerator);
 
-		const { line: margins } = Configuration.charts.margin;
-		const circleRadius = this.getCircleRadius();
-		gLines.selectAll("circle.dot")
-			.data(function(d, i) {
-				const parentDatum = select(this).datum() as any;
-
-				return self.addLabelsToDataPoints(parentDatum, i);
-			})
-			.transition(transitionToUse)
-			.attr("cx", d => this.x(d.label) + this.x.step() / 2)
-			.attr("cy", d => this.y(d.value))
-			.attr("r", circleRadius)
-			.attr("fill", d => this.getCircleFill(circleRadius, d))
-			.attr("stroke", d => this.colorScale[d.datasetLabel](d.label));
-	}
-
-	resizeChart() {
-		const chartSize: any = this.getChartSize(this.container);
-		const dimensionToUseForScale = Math.min(chartSize.width, chartSize.height);
-
-		// Resize the SVG
-		select(this.holder).select("svg")
-				.attr("width", `${dimensionToUseForScale}px`)
-				.attr("height", `${dimensionToUseForScale}px`);
-
-		this.updateXandYGrid(true);
-		// Scale out the domains
-		this.setXScale();
-		this.setYScale();
-
-		// Set the x & y axis as well as their labels
-		this.setXAxis(true);
-		this.setYAxis(true);
-
-		this.updateElements(false, null);
-
-		super.resizeChart();
-	}
-
-	setXScale () {
-		super.setXScale();
-
-		this.x.padding(0); // override BaseAxisChart padding so points aren't misaligned by a few pixels
-	}
-
-	resetOpacity() {
-		const circleRadius = this.getCircleRadius();
-		this.innerWrap.selectAll("circle")
-			.attr("stroke-opacity", Configuration.charts.resetOpacity.opacity)
-			.attr("fill", d => this.getCircleFill(circleRadius, d));
-	}
-
-	reduceOpacity(exception) {
-		const circleRadius = this.getCircleRadius();
-		select(exception).attr("fill-opacity", false);
-		select(exception).attr("stroke-opacity", Configuration.charts.reduceOpacity.opacity);
-		select(exception).attr("fill", (d: any) => this.getCircleFill(circleRadius, d));
-	}
-
-	addDataPointEventListener() {
-		const self = this;
-		const { accessibility } = this.options;
-
-		this.svg.selectAll("circle.dot")
-			.on("click", function(d) {
-				self.dispatchEvent("line-onClick", d);
-			})
-			.on("mouseover", function(d) {
-				select(this)
-					.attr("stroke-width", Configuration.lines.points.mouseover.strokeWidth)
-					.attr("stroke", self.colorScale[d.datasetLabel](d.label))
-					.attr("stroke-opacity", Configuration.lines.points.mouseover.strokeOpacity);
-
-				self.showTooltip(d, this);
-				self.reduceOpacity(this);
-			})
-			.on("mousemove", function(d) {
-				const tooltipRef = select(self.holder).select("div.chart-tooltip");
-
-				const relativeMousePosition = mouse(self.holder as HTMLElement);
-				tooltipRef.style("left", relativeMousePosition[0] + Configuration.tooltip.magicLeft2 + "px")
-					.style("top", relativeMousePosition[1] + "px");
-			})
-			.on("mouseout", function(d) {
-				const { strokeWidth, strokeWidthAccessible } = Configuration.lines.points.mouseout;
-				select(this)
-					.attr("stroke-width", accessibility ? strokeWidthAccessible : strokeWidth)
-					.attr("stroke", self.colorScale[d.datasetLabel](d.label))
-					.attr("stroke-opacity", Configuration.lines.points.mouseout.strokeOpacity);
-
-				self.hideTooltip();
-			});
+		super.updateElements(animate);
 	}
 }